Remove redundant `ControlItemText` class
